requirements.txt
datasets 
transformers 
sentencepiece 
torch
pytorch-lightning
pyyaml
python-dotenv
langchain
tiktoken
faiss-cpu
gradio
chromadb
huggingface_hub
llama-cpp-python